+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - #114 Adds "How To Use Koffee" link in settings window
 - #70 Changes undo move/copy to undo individual file operations for more correct behavior when folders were merged
 - #123 Fixes path suggestions not being sorted by MRU
+- #126 Fixes search results to sort by current sort setting instead of always name ascending
 - #118 Fixes window size and position on opening to fit inside the current screen
 - Simplifies same-folder copy naming
 - Reorders keybindings in Settings list to hopefully be more logical